Factor bits out of PGExtras
